feat: add the function to get time from network and convert it in a global variable
