+++ RTC_Serial_Modem/platformio.ini
@@ -12,3 +12,4 @@
 platform = espressif32
 board = esp-wrover-kit
 framework = arduino
+lib_deps = plerup/EspSoftwareSerial@^8.2.0
