+++ RTC_Serial_Modem/src/features/NetworkTimeConverter.hpp
@@ -1,0 +1,5 @@
+#include <Arduino.h>
+#include <iostream> // std::cout
+#include <string>   // std::string, std::stoi
+
+void convertNetworkTime2Integer();
