+++ RTC_Serial_Modem/src/features/global_variables.cpp
@@ -1,0 +1,18 @@
+#include "global_variables.hpp"
+#include <Arduino.h>
+/*
+int NetworkTime::networkhour;
+int NetworkTime::networkminutes;
+int NetworkTime::networksenconds;
+
+int NetworkDate::networkyear;
+int NetworkDate::networkmonth;
+int NetworkDate::networkday;
+*/
+
+int NETWORK_HOUR = 0;
+int NETWORK_MINUTE= 0;
+int NETWORK_SECOND= 0;
+int NETWORK_YEAR= 0;
+int NETWORK_MONTH= 0;
+int NETWORK_DAY= 0;
